test: split browser address reuse coverage into its own suite
